feat: Add new notification center layouts and routes and page to configure account and pref

+++ backend/src/modules/m-user/m-user.controller.ts
@@ -1,26 +1,26 @@
+import { Auth, GetUser } from '@/auth/auth.decorator';
 import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   Param,
-  Patch,
   Post,
+  Put,
 } from '@nestjs/common';
 import { ApiParam, ApiTags } from '@nestjs/swagger';
 import * as prisma from '@prisma/client';
 import { AccountsService } from './accounts/accounts.service';
 import { UpsertAccountDto } from './accounts/dto/upsert-account.dto';
 import { Account } from './accounts/entities/account.entity';
-import { CreateUserPreferenceDto } from './user-preferences/dto/create-user-preference.dto';
-import { UpdateUserPreferenceDto } from './user-preferences/dto/update-user-preference.dto';
+import { UpsertUserPreferenceDto } from './user-preferences/dto/upsert-user-preference.dto';
 import { UserPreference } from './user-preferences/entities/user-preference.entity';
 import { UserPreferencesService } from './user-preferences/user-preferences.service';
 import { CreateUserDto } from './users/dto/create-user.dto';
-import { UpdateUserDto } from './users/dto/update-user.dto';
+import { VerifyUserDto } from './users/dto/verify-user.dto';
 import { User } from './users/entities/user.entity';
 import { UsersService } from './users/users.service';
-import { VerifyUserDto } from './users/dto/verify-user.dto';
 
 @ApiTags('User Module')
 @Controller('m-user')
@@ -50,14 +50,6 @@ export class MUserController {
     return this.usersService.findAll();
   }
 
-  @Patch('users/:userId')
-  updateUser(
-    @Param('userId') userId: string,
-    @Body() updateUserDto: UpdateUserDto,
-  ): Promise<User> {
-    return this.usersService.update(userId, updateUserDto);
-  }
-
   /**
    * Accounts
    **/
@@ -66,36 +58,29 @@ export class MUserController {
     name: 'channelType',
     enum: prisma.$Enums.ChannelType,
   })
-  @Post('users/:userId/channel/:channelType/accounts')
-  createAccount(
+  @Auth()
+  @Put('users/:userId/channel/:channelType/accounts')
+  upsertAccount(
+    @GetUser() user: User,
     @Param('userId') userId: string,
     @Param('channelType') channelType: prisma.$Enums.ChannelType,
     @Body() upsertAccountDto: UpsertAccountDto,
   ): Promise<Account> {
+    if (user.id !== userId) {
+      throw new ForbiddenException();
+    }
     return this.accountsService.upsert(userId, channelType, upsertAccountDto);
   }
 
-  @ApiParam({
-    name: 'channelType',
-    enum: prisma.$Enums.ChannelType,
-  })
-  @Get('users/:userId/channel/:channelType/accounts')
-  findOneAccount(
-    @Param('userId') userId: string,
-    @Param('channelType') channelType: prisma.$Enums.ChannelType,
-  ): Promise<Account> {
-    return this.accountsService.findOne(userId, channelType);
-  }
-
-  @Get('accounts')
-  findAllAccounts(): Promise<Account[]> {
-    return this.accountsService.findAll();
-  }
-
+  @Auth()
   @Get('users/:userId/accounts')
   findAllUserAccountsByUserId(
+    @GetUser() user: User,
     @Param('userId') userId: string,
   ): Promise<Account[]> {
+    if (user.id !== userId) {
+      throw new ForbiddenException();
+    }
     return this.accountsService.findAllByUserId(userId);
   }
 
@@ -103,43 +88,34 @@ export class MUserController {
     name: 'channelType',
     enum: prisma.$Enums.ChannelType,
   })
-  @Patch('users/:userId/channel/:channelType/accounts')
-  updateAccount(
-    @Param('userId') userId: string,
-    @Param('channelType') channelType: prisma.$Enums.ChannelType,
-    @Body() upsertAccountDto: UpsertAccountDto,
-  ): Promise<Account> {
-    return this.accountsService.upsert(userId, channelType, upsertAccountDto);
-  }
-
-  @ApiParam({
-    name: 'channelType',
-    enum: prisma.$Enums.ChannelType,
-  })
+  @Auth()
   @Delete('users/:userId/channel/:channelType/accounts')
   removeAccount(
+    @GetUser() user: User,
     @Param('userId') userId: string,
     @Param('channelType') channelType: prisma.$Enums.ChannelType,
   ) {
+    if (user.id !== userId) {
+      throw new ForbiddenException();
+    }
     return this.accountsService.remove(userId, channelType);
   }
-
-  // async verify(
-  //   userId: string,
-  //   channelType: prisma.$Enums.ChannelType,
-  //   otpCode: string,
-  // ) {
 
   @ApiParam({
     name: 'channelType',
     enum: prisma.$Enums.ChannelType,
   })
+  @Auth()
   @Post('users/:userId/channel/:channelType/accounts/verify')
   verifyAccount(
+    @GetUser() user: User,
     @Param('userId') userId: string,
     @Param('channelType') channelType: prisma.$Enums.ChannelType,
     @Body() verifyUserDto: VerifyUserDto,
   ) {
+    if (user.id !== userId) {
+      throw new ForbiddenException();
+    }
     return this.accountsService.verify(userId, channelType, verifyUserDto);
   }
 
@@ -151,16 +127,21 @@ export class MUserController {
     name: 'channelType',
     enum: prisma.$Enums.ChannelType,
   })
-  @Post(
+  @Auth()
+  @Put(
     'users/:userId/channel/:channelType/notification-categories/:notificationCategoryId/user-preferences',
   )
-  createUserPreference(
+  upsertUserPreference(
+    @GetUser() user: User,
     @Param('userId') userId: string,
     @Param('channelType') channelType: prisma.$Enums.ChannelType,
     @Param('notificationCategoryId') notificationCategoryId: string,
-    @Body() createUserPreferenceDto: CreateUserPreferenceDto,
+    @Body() createUserPreferenceDto: UpsertUserPreferenceDto,
   ): Promise<UserPreference> {
-    return this.userPreferencesService.create(
+    if (user.id !== userId) {
+      throw new ForbiddenException();
+    }
+    return this.userPreferencesService.upsert(
       userId,
       notificationCategoryId,
       channelType,
@@ -168,55 +149,37 @@ export class MUserController {
     );
   }
 
-  @Get('user-preferences')
-  findAllUserPreferences(): Promise<UserPreference[]> {
-    return this.userPreferencesService.findAll();
+  @ApiParam({
+    name: 'channelType',
+    enum: prisma.$Enums.ChannelType,
+  })
+  @Auth()
+  @Put('users/:userId/channel/:channelType/user-preferences')
+  updateManyUserPreferences(
+    @GetUser() user: User,
+    @Param('userId') userId: string,
+    @Param('channelType') channelType: prisma.$Enums.ChannelType,
+    @Body() upsertUserPreferenceDto: UpsertUserPreferenceDto,
+  ) {
+    if (user.id !== userId) {
+      throw new ForbiddenException();
+    }
+    return this.userPreferencesService.updateManyByUserIdAndChannelType(
+      userId,
+      channelType,
+      upsertUserPreferenceDto,
+    );
   }
 
+  @Auth()
   @Get('users/:userId/user-preferences')
   findAllUserPreferencesByUserId(
+    @GetUser() user: User,
     @Param('userId') userId: string,
   ): Promise<UserPreference[]> {
+    if (user.id !== userId) {
+      throw new ForbiddenException();
+    }
     return this.userPreferencesService.findAllByUserId(userId);
-  }
-
-  @ApiParam({
-    name: 'channelType',
-    enum: prisma.$Enums.ChannelType,
-  })
-  @Get(
-    'users/:userId/channel/:channelType/notification-categories/:notificationCategoryId/user-preferences',
-  )
-  findOneUserPreference(
-    @Param('userId') userId: string,
-    @Param('channelType') channelType: prisma.$Enums.ChannelType,
-    @Param('notificationCategoryId') notificationCategoryId: string,
-  ): Promise<UserPreference> {
-    return this.userPreferencesService.findOne(
-      userId,
-      notificationCategoryId,
-      channelType,
-    );
-  }
-
-  @ApiParam({
-    name: 'channelType',
-    enum: prisma.$Enums.ChannelType,
-  })
-  @Patch(
-    'users/:userId/channel/:channelType/notification-categories/:notificationCategoryId/user-preferences',
-  )
-  updateUserPreference(
-    @Param('userId') userId: string,
-    @Param('channelType') channelType: prisma.$Enums.ChannelType,
-    @Param('notificationCategoryId') notificationCategoryId: string,
-    @Body() updateUserPreferenceDto: UpdateUserPreferenceDto,
-  ): Promise<UserPreference> {
-    return this.userPreferencesService.update(
-      userId,
-      notificationCategoryId,
-      channelType,
-      updateUserPreferenceDto,
-    );
   }
 }
